+++ BambooTracker/bamboo_tracker.hpp
@@ -187,7 +187,7 @@ public:
 	// Export
 	bool exportToWav(std::string file, int loopCnt, std::function<bool()> f);
 	bool exportToVgm(std::string file, int target, bool gd3TagEnabled, GD3Tag tag, std::function<bool()> f);
-	bool exportToS98(std::string file, int target, bool tagEnabled, S98Tag tag, std::function<bool()> f);
+	bool exportToS98(std::string file, int target, bool tagEnabled, S98Tag tag, int rate, std::function<bool()> f);
 
 	// Backup
 	void backupModule(std::string file);